scripts/requirements.txt
python-dotenv>=1.0
requests>=2.31
anthropic>=0.40
feedparser>=6.0
python-telegram-bot>=21.0
